ccviR: Calculate the NatureServe Climate Change Vulnerability Index in R

The ccviR package implements the NatureServe Climate Change Vulnerability Index (CCVI) version 3.02 https://www.natureserve.org/conservation-tools/climate-change-vulnerability-index in an R package and Shiny App. The package allows all of the geospatial aspects of calculating the CCVI to be done in R, removing the need for separate GIS calculations. It also enables easier sensitivity analysis of the Index. The app provides an interactive application designed to offer a user-friendly and simple interface for calculating the NatureServe CCVI. See Young et. al (2012) doi:10.7208/9780226074641-007 and Young et. al. (2015) doi:10.1002/wsb.478 for a detailed description of the index.

See also

Author

Maintainer: Sarah Endicott sarah.endicott@canada.ca (ORCID)

Authors:

Ilona Naujokaitis-Lewis

Other contributors:

Steffi LaZerte (ORCID) [contributor]

His Majesty the King in Right of Canada, as represented by the Minister of Environment and Climate Change [copyright holder]