Load climate variables

Load climate variables and store them in a list that can be used with analyze_spatial. The climate variables should first be prepared using prep_clim_data.

get_clim_vars(root_pth, scenario_names = "scn1")

Arguments

root_pth

A folder location where all the climate data is stored. The names must match one of c("MAT.*tif$", "CMD.*tif$", "clim_poly.*shp", "MAP.*tif$", "ccei.*tif$|CCEI.*tif$","MWMT.*tif$|HTN.*tif$") and the first three are required.

scenario_names

character vector with names that identify multiple future climate scenarios. If this is supplied the raster file must include the scenario name as a suffix to the pattern mentioned above eg. if there are two MAT files "MAT_RCP 4.5.tif" and "MAT_RCP 8.5.tif" the scenario names should be "RCP 4.5" and "RCP 8.5". This will happen automatically if the scenario name is provided to prep_clim_data.

Value

A list of climate variables with names "mat", "cmd", "map", "ccei", "htn", "clim_poly". If multiple scenarios are used mat, cmd and ccei will be RasterStacks with one layer per scenario.

Examples

pth <- system.file("extData/clim_files/processed", package = "ccviR")

# scenario names
scn_nms <- c("RCP 4.5", "RCP 8.5")

get_clim_vars(pth, scn_nms)
#> Error: directory  does not exist