Run the spatial analysis

Run the required spatial analysis to create the spat_df input for calc_vulnerability and clip the range polygon to the appropriate scales.

run_spatial(
  range_poly,
  scale_poly,
  clim_vars_lst,
  non_breed_poly = NULL,
  ptn_poly = NULL,
  hs_rast = NULL,
  hs_rcl = NULL,
  gain_mod = 1
)

Arguments

range_poly

an sf polygon object giving the species range.

scale_poly

an sf polygon object giving the area of the assessment

clim_vars_lst

a list of climate data, the result of get_clim_vars

non_breed_poly

Optional. An sf polygon object giving the species range in the non-breeding season.

ptn_poly

Optional. An sf polygon object giving the locations that are considered part of the physiological thermal niche (See NatureServe Guidelines for definition).

hs_rast

Optional. A raster with results from a model of the change in habitat suitability caused by climate change.

hs_rcl

a matrix used to classify hs_rast into 0: not suitable, 1: lost, 2: maintained, 3: gained. See reclassify for details on the matrix format

Value

a list with elements: spat_table the results of the spatial analysis, range_poly_assess the range polygon clipped to the assessment area, and range_poly_clim the range polygon clipped to the extent of the climate data. spat_table contains the following columns:

MAT_#

The percentage of the species' range that is exposed to each class of change in mean annual temperature between the historical normal and predicted climate. Class 1 has the highest exposure and Class 6 the lowest

CMD_#

The percentage of the species' range that is exposed to each class of change in climate moisture deficit between the historical normal and predicted climate. Class 1 has the highest exposure and Class 6 the lowest

CCEI_#

The percentage of the species' non-breeding range that falls into each climate change exposure index class. Class 4 indicates high exposure while Class 1 indicates low exposure

perc_non_breed_not_over_ccei

The precentage of the non-breeding range that does not overlap with the CCEI raster data

HTN_#

The percentage of the species' range that is exposed to each class of variation between the historical coldest and warmest monts. Class 1 has the smallest variation and Class 4 is the largest

PTN

The percentage of the species' range that falls into cool or cold environments that may be lost or reduced in the assessment area as a result of climate change

MAP_max/min

The maximum and minimum historical mean annual precipitation in the species' range

range_change

The projected decrease in range size as a percentage of current range size. Negative numbers indicate an increase in range size

range_overlap

The percentage of the current range that is projected to remain in the future range.

range_size

The area of the species' range in m2

Details

The range polygon will be clipped to the area overlapping the scale_poly and also to the area overlapping the extent of the climate data polygon. The range within the assessment area is used to calculate all results except the historical thermal and hydrological niches for which the range within the extent of the climate data is used.

Examples

base_pth <- system.file("extdata", package = "ccviR")

clim_vars <- get_clim_vars(file.path(base_pth, "clim_files/processed"))

run_spatial(
  range_poly = sf::read_sf(file.path(base_pth, "rng_poly_high.shp"), agr = "constant"),
  scale_poly = sf::read_sf(file.path(base_pth, "assess_poly.shp"), agr = "constant"),
  clim_vars_lst = clim_vars,
  hs_rast = raster::raster(file.path(base_pth, "HS_rast_high.tif")),
  hs_rcl = matrix(c(0:7, 0, 1, 2, 2 ,2, 2, 2, 3), ncol = 2)
)
#> performing spatial analysis
#> $spat_table
#> # A tibble: 1 × 27
#>   MAT_1 MAT_2 MAT_3 MAT_4 MAT_5 MAT_6 CMD_1 CMD_2 CMD_3 CMD_4 CMD_5 CMD_6 CCEI_1
#>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1    10    90     0     0     0     0    10    90     0     0     0     0     NA
#> # … with 14 more variables: CCEI_2 <dbl>, CCEI_3 <dbl>, CCEI_4 <dbl>,
#> #   perc_non_breed_not_over_ccei <dbl>, HTN_1 <dbl>, HTN_2 <dbl>, HTN_3 <dbl>,
#> #   HTN_4 <dbl>, PTN <dbl>, MAP_max <dbl>, MAP_min <dbl>, range_change <dbl>,
#> #   range_overlap <dbl>, range_size <dbl>
#> 
#> $range_poly_assess
#> Simple feature collection with 1 feature and 2 fields
#> Attribute-geometry relationship: 2 constant, 0 aggregate, 0 identity
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 0 ymin: 750 xmax: 1000 ymax: 1000
#> Projected CRS: NAD83(CSRS) / Ontario MNR Lambert
#> # A tibble: 1 × 3
#>     FID FID.1                                      geometry
#> * <dbl> <dbl>                                 <POLYGON [m]>
#> 1     0     0 ((0 750, 0 1000, 1000 1000, 1000 750, 0 750))
#> 
#> $range_poly_clim
#> Simple feature collection with 1 feature and 2 fields
#> Attribute-geometry relationship: 2 constant, 0 aggregate, 0 identity
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 0 ymin: 750 xmax: 1000 ymax: 1000
#> Projected CRS: NAD83(CSRS) / Ontario MNR Lambert
#> # A tibble: 1 × 3
#>     FID FID.1                                      geometry
#> * <dbl> <dbl>                                 <POLYGON [m]>
#> 1     0     0 ((0 750, 0 1000, 1000 1000, 1000 750, 0 750))
#>