Calculate the vulnerability index score

Calculate the vulnerability index score from the completed inputs.

calc_vulnerability(spat_df, vuln_df, tax_grp, n_rnds = 1000)

Arguments

spat_df

data.frame of spatial inputs produced by run_spatial

vuln_df

data.frame of answers to vulnerability questions. The format must match make_vuln_df.

tax_grp

taxonomic group of the species. Must match one of the options below

n_rnds

number of monte carlo repetitions for determining variability when multiple values are given for a factor

Value

A list:

index

The index value as a string. EV: Extremely Vulnerable, HV: Highly Vulnerable, MV: Moderately Vulnerable, LV: Less Vulnerable, IE: Insufficient Evidence

conf_index

An index of the confidence in the index based on the monte carlo. Options are Low, Moderate, High, Very High or Insufficient Evidence.

index_conf

A data.frame with the proportion of monte carlo runs with each index value

mig_exp

An index of the migratory exposure. Options are Low, Moderate, High, or "N/A"

d_score

Total score from section D the modelled response to climate change section

b_c_score

Total score from sections B and C the indirect exposure, sensitivity and adaptive capacity sections

vuln_df

Answers to each question plus any comments. Numerical scores correspond to: -1) Unknown, 0) Neutral, 1) Somewhat increase, 2) Increase, 3) Greatly increase. If there are 2 numbers then two boxes were checked and the average of the scores was used in the total score

n_b_factors

Number of factors scored in section B

n_c_factors

Number of factors scored in section C

n_d_factors

Number of factors scored in section D

slr_vuln

Whether the species' vulnerability to sea level rise and dispersal limitations led to an index of EV

Details

Options for tax_grp are Vascular Plant, Nonvascular Plant, Lichen, Invert-Insect, Invert-Mollusk, Invert-Other, Fish, Amphibian, Reptile, Mammal, Bird

Examples

base_pth <- system.file("extData", package = "ccviR")

clim_vars <- get_clim_vars(file.path(base_pth, "clim_files/processed"))
#> Error: directory /clim_files/processed does not exist

spat_res <- run_spatial(
  range_poly = sf::read_sf(file.path(base_pth, "rng_poly_high.shp"), agr = "constant"),
  scale_poly = sf::read_sf(file.path(base_pth, "assess_poly.shp"), agr = "constant"),
  clim_vars_lst = clim_vars,
  hs_rast = raster::raster(file.path(base_pth, "HS_rast_high.tif")),
  hs_rcl = matrix(c(0:7, 0, 1, 2, 2 ,2, 2, 2, 3), ncol = 2)
)
#> performing spatial analysis
#> Error in setdiff(names(clim_vars_lst), c("mat", "cmd", "map", "ccei",     "htn", "clim_poly")): object 'clim_vars' not found

# vulnerability factor table with score 1 (somewhat increase vulnerability)
# for all factors
vuln <- make_vuln_df("test_species", val1 = 1, mig = 1)

calc_vulnerability(spat_res$spat_table, vuln, "Bird")
#> calculating vulnerability index
#> Error in as.vector(y): object 'spat_res' not found