Prepare climate data

Prepare data from raw to form needed for calculating the index. See the NatureServe Guidelines for details on how the data is prepared.

run_prep_data(
  mat_norm,
  mat_fut,
  cmd_norm,
  cmd_fut,
  ccei = NULL,
  map = NULL,
  mwmt = NULL,
  mcmt = NULL,
  clim_poly = NULL,
  in_folder = NULL,
  out_folder,
  reproject = FALSE,
  overwrite = FALSE
)

Arguments

mat_norm, mat_fut, cmd_norm, cmd_fut, ccei, map, mwmt, mcmt, clim_poly

file paths to find data if in_folder is not given

in_folder

file path where files are stored. Files must be named according to the convention described in details

out_folder

file path where the processed files will be saved

reproject

should the data be re-projected to lat/long? Not recommended.

overwrite

should existing files in out_folder be overwritten?

Value

Returns nothing. Processed data is saved in out_folder

Details

Definition of input data sets and file names required in in_folder:

mat_norm:

"MAT" mean annual temperature for the historical normal period

mat_fut:

"MAT_2050" mean annual temperature for the future under climate change it can be any number eg 2050, 2100

cmd_norm:

"CMD" climate moisture deficit for the historical normal period

cmd_fut:

"CMD_2050" climate moisture deficit for the future under climate change it can be any number eg 2050, 2100

ccei:

"CCEI" Climate Change Exposure Index from NatureServe website

map:

"MAP" mean annual precipitation for the historical normal period

mwmt:

"MWMT" mean warmest month temperature for the historical normal period

mcmt:

"MCMT" mean coldest month temperature for the historical normal period

clim_poly:

An optional shapefile with a polygon of the range of the climate data. It will be created from the climate data if it is missing but it is faster to provide it.

Accepted raster file types are ".asc", ".tif", ".nc", ".grd" and ".img"

See also

Examples

pth_in <- system.file("extData/clim_files/raw", package = "ccviR")

pth_out <- system.file("extData/clim_files/processed", package = "ccviR")

run_prep_data(in_folder = pth_in, out_folder = pth_out, overwrite = TRUE)
#> Error: out_folder does not exist